pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
stripe = "^11.3.0"
celery = "^5.4.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
django-celery-beat = "^2.7.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
